Switch from Laravel Elixir to Laravel Mix
Even though Laravel Elixir was recently updated to its latest version
on this project it is a EOL module. The new tool, specially for those
familiar to Laravel, is Laravel Mix.

With this Pull Request it's providen a full working version of Jigsaw
using Laravel Mix instead of Laravel Elixir.

It's important to mention that Mix surpasses Elixir, providing much
more functionality. The only exception to this is: `.exec`. Thus we
use a Webpack plugin that is able to call external commands so we
are able to call Jigsaw. The call to Jigsaw still complies with
issue #85, where it is called from local or global composer,
whichever is available.

Now, `npm run webpack` is the equivalent of running `gulp` and `npm
run dev` is the one for `gulp watch`. Both of them run for local
builds (`--env=local`). To run the production version we can use
`npm run production`. There are other commands as well, coming
from Mix, all of them on `package.json`.